refactor: remove  util, and integrate with
