+++ src/utils/allIconInfo.ts
@@ -1,7 +1,7 @@
 import type { SimpleIcon } from 'simple-icons';
 import * as icons from 'simple-icons';
 
-import { customIconArray } from './customIconObjects';
+import { customIconArray, customIconObject } from './customIconObjects';
 
 export const makeIconInfoArray = () => {
   const iconArr: SimpleIcon[] = Object.values(icons);
@@ -9,4 +9,12 @@ export const makeIconInfoArray = () => {
 
   const allIconArr = [...iconArr, ...customIconArr];
   return allIconArr;
+};
+
+export const getIconDetail = (target: string) => {
+  const allIconArr = { ...icons, ...customIconObject };
+
+  const iconKey = ('si' + target) as keyof typeof allIconArr;
+
+  return allIconArr[iconKey];
 };
